+++ tests/evm_suite_test.go
@@ -38,7 +38,7 @@ var (
 )
 
 type EvmTestSuite struct {
-	TestSuite
+	*TestSuite
 	ethPrivKey cryptotypes.PrivKey
 }