+++ CHANGELOG.md
@@ -41,6 +41,9 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 * Add query oracle reward in the crosschain module
 * Check fxcored version when synchronizing blocks from scratch
+* Add denom many to one support
+* Update RegisterCoinProposal support denom many to one
+* Add UpdateDenomAliasProposal and MsgConvertDenom
 
 ### Improvements